Move config values from Utils to Config
