+++ love/src/components/CSCSummary/CSCSummary.jsx
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import styles from './CSCSummary.module.css';
 import CSCRealm from './CSCRealm/CSCRealm';
 import Panel from '../GeneralPurpose/Panel/Panel';
-import ManagerInterface, { hasFakeData } from '../../Utils';
-import { CSCSummaryHierarchy } from '../../Config';
+import ManagerInterface from '../../Utils';
+import { hasFakeData, CSCSummaryHierarchy } from '../../Config';
 
 export default class CSCSummary extends Component {
   constructor(props) {
