+++ love/src/components/ScriptQueue/ScriptQueue.jsx
@@ -8,7 +8,8 @@ import DraggableScript from './Scripts/DraggableScript/DraggableScript';
 import styles from './ScriptQueue.module.css';
 import Panel from '../GeneralPurpose/Panel/Panel';
 import StatusText from '../GeneralPurpose/StatusText/StatusText';
-import ManagerInterface, { hasCommandPrivileges, hasFakeData } from '../../Utils';
+import ManagerInterface from '../../Utils';
+import { hasCommandPrivileges, hasFakeData } from '../../Config';
 
 /**
  * Display lists of scripts from the ScriptQueue SAL object. It includes: Available scripts list, Waiting scripts list and Finished scripts list.
