+++ love/src/components/ScriptQueue/Scripts/WaitingScript/WaitingScript.jsx
@@ -7,7 +7,7 @@ import ScriptStatus from '../../ScriptStatus/ScriptStatus';
 import { getStatusStyle } from '../Scripts';
 import UploadButton from '../../../HealthStatusSummary/Button/UploadButton';
 import Button from '../../../GeneralPurpose/Button/Button';
-import { hasCommandPrivileges } from '../../../../Utils';
+import { hasCommandPrivileges } from '../../../../Config';
 import HeartbeatIcon from '../../../icons/HeartbeatIcon/HeartbeatIcon';
 
 export default class WaitingScript extends Component {
